+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.6] - 20 Feb 2021
+* Convert filter list package to generic list filter package
+* `allTextList` changed to `listData`
+* `selectedTextList` changed to `selectedListData`
+* `FilterListWidget` and `FilterListDialog` can filter any type if list
+* Added `validateSelectedItem` callback to validate which item needs to be selected 
+* Added `onItemSearch` callback to expose search mechanism on user side to filter list.'
 ## [0.0.5] - 22 Sep 2020
 * Add `FilterListWidget` widget.
 * `FilterList.showFilterList` is renamed to `FilterListDialog.display`
@@ -9,6 +16,7 @@
 
 ## [0.0.3] - 02 Mar 2020
 * Added pop-up height
+
 * Added pop-up width
 * Added header hide option
 * Added search field hide option